LPS-114374 | LPS-111958 Move reusable code to utils
